requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.1
flask==3.0.0
flask-cors==4.0.0

